fix(estate): remove duplicate 'state' field in property form view
Removed the redundant 'state' field from the form view to avoid adding
the same field multiple times, which is not supported by Odoo views. The
'state' field is now only displayed in the header as a statusbar.

Warning: Same field multiple times in a view. Adding a field only once
to a list or form view is supported by Odoo. Adding it multiple times
can cause rendering issues.

Refer:
https://www.odoo.com/documentation/17.0/developer/tutorials/server_framework_101/11_sprinkles.html